Don't build $_POST into querystring (see below).
This utility function is used to generate the given querystring, except any
internally used variables, such as __path__ for web server -> da.gd pass-off.
However, the only place it is currently used is at the entry point (index.php)
and only when we are internally routing to a redirect. In this case, passing
POST arguments as GET arguments to the redirect is the *incorrect* thing to
do anyway, so nothing should have ever relied on this.