Add an explicit `Expires` header.
On a captive portal wifi, I was able to trigger that I could
successfully load /ip in a browser before passing the captive
portal.

This is less of a dagd bug, but in an effort to combat misbehaving
proxies, we now add an explicit `Expires: -1` header.

(As an aside, the captive portal I triggered this on seems really
weird, but I think the repro is something like this:

When logged in, go to /ip. Then log out (e.g. by spoofing a MAC address
and reconnecting). On the first request to /ip, you'll actually be able
to hit /ip (the request actually shows up in the access log, so it's not
just cached via the captive portal proxy). But then if you hit it again,
you'll hit the captive portal auth page and be given a set of headers
that indicate a squid cache MISS.

I'm not sure why this is or if `Expires: -1` will fix it, but this is an
attempt. The reason we care is so that people can hit up da.gd to
legitimately trigger the captive portal login page if they wish without
having to go to another page because da.gd loads for some unexplainable
reason).

Signed-off-by: Ricky Elrod <ricky@elrod.me>