Allow for querying multiple generic-TLD hosts...
In the same way that the OpenBSD whois client works, we now (with the
default configuration) query multiple TLD-based whois servers. In the
default configuration we first try TLD.whois-servers.net, then
whois.nic.TLD. This is the reverse of the OpenBSD client, but with the
goal of not changing how existing responses worked before this commit.
Everything that worked in the past should be exactly the same, but now
we should be able to handle new (2003+) gTLDs.

Signed-off-by: Ricky Elrod <ricky@elrod.me>