[whois] Fix bug in referral logic (see below)
In the case where we are referred to a target server in our parsing,
we would stop parsing and attempt to hit the new server.

In the case where we couldn't reach the new server, we would rightly
bail out of that attempt and return the transient result. The only
problem is that in that case, we never finished parsing (nor storing)
the transient server result.

The net result is that in this case, we'd end up rendering the first
few lines of transient output and nothing more - a most nonsensical
thing to do.

After this commit, we store the entire transient output. If we find
that the referral server is blacklisted, or is not blacklisted but
can't be reached, we show the transient result. Otherwise, we show the
detailed result -- the one the referral target server gives us.

Signed-off-by: Ricky Elrod <ricky@elrod.me>