[shorten] Store a shasum of the long url
Use the shasum for indexing so that we can more
quickly check if a randomized shorturl has already
been assigned to a randomized-shorturl request for
some long url.

Since the `longurl` field is `TEXT` and there
are some legitimate longurls stored in the production
database that go beyond the InnoDB 3076 key length
anyway, simply converting `longurl` to `varchar` and
adding that to the index is not sufficient.

Signed-off-by: Ricky Elrod <ricky@elrod.me>