Avoid __path__ hack for routing, use REQUEST_URI
Apache normalizes paths sometimes and we want to avoid that so that
things like /headers/http://google.com work. When Apache normalizes it,
it removes the second slash there.

Really the slashes should be encoded, but we want this to work from a
browser without the user needing to care about encoding.

Signed-off-by: Rick Elrod <rick@elrod.me>