[experimental] add infra for post-request actions
This adds infrastructure for running code (specified by a callback) after a
request has been handled and its response provided and connection closed.

Controllers can register callbacks, and the base controller will re-register
them with the response object within its finalize() method. We likely want to
rework this before going live (so this should be considered a proof of concept)
as some controllers and applications override finalize() and likely won't call
the supermethod when they do.

Other issues with this approach:
- We are limited by the PHP interpreter's timeout (often 30s by default, but
  varies by distro/hosting provider/etc.).
- We could run into issues where long-running post-processors tie up FPM (or
  similar) workers which drastically limits the number of requests that we can
  serve at any given moment.

In spite of these issues, this infrasturcture might still be useful for
occasional/rare actions, such as administrative actions on endpoints run by
site administrators infrequently.

Endpoints which are hit frequently and have post-processors that might not
return immediately should probably use an external task queue for these types
of actions. First-party support for integration with such a task queue might
be implemented in the future if the need arises. There is precedent for
integration with external services like this within the caching subsystem
where memcache support is first-party (though not widely tested nor used in
the production site).

Basically: Don't use this on heavily-trafficked endpoints, especially with
post-processors that take time to return.

Signed-off-by: Rick Elrod <rick@elrod.me>