Introduce DaGdTask/DaGdWorker and CLI for worker
This introduces a simple task queue using a MySQL table. This acts as a
functional replacement for the 'postprocess' branch (never merged into
master) and allows callers to register tasks to queue.

Tasks are subclasses of DaGdTask which implement a run() method.
A singular parameter (of any serializable PHP type) can optionally be
stored with the task and will be sent to the task's run() when the task
is popped from the queue by the worker. Tasks can also optionally store
their result (again automatically serialized by PHP's serialize()
function).

Currently workers are single-threaded, but do fork child processes for
running a task. When a worker claims a task, it will fork an
intermediate child process, which then forks either one or two more
children, depending on whether or not the task requested a timeout. (The
non-task-running fork simply sleeps and acts a timer to know when to
kill the fork doing the actual work).

In theory, we could run many tasks on a worker at once, but currently
the implementation waits for a single task to complete (or time out)
before selecting a new task if one is available. However, many workers
can safely be spawned alongside each other. I have tested with 35 with
no issues.

When a task is complete, it will establish a new connection to the
database, to ensure that the previouly-shared database handle hasn't
been closed when a previous task's child processes ended. When the child
process ends, the worker-global database handle also gets reconnected
(out of necessity, it will get closed when the child process exits) and
this handle is used to loop once per second looking for new work.

When a worker looks for work it selects for update (`SELECT .. FOR
UPDATE`) to prevent another worker from selecting the same task.

Currently there is no concept of task-reaping. If a worker dies after
claiming a task, but before marking it as completed, there is no
mechanism to handle that.

There is also no built-in retry logic. If a task fails, it is marked as
failed and that is it. This could feasibly be added in the future.

Signed-off-by: Rick Elrod <rick@elrod.me>