DaGdController#execute can return DaGdResponse
Allow DaGdController#execute to return a DaGdResponse, and in that
case, that response is expected to rule the remainder of the response
cycle.

This is a somewhat risky change, but the net result should be that
controllers opting to use execute() can return error responses that
obey the normal response type selection logic. e.g. just because we
are in execute() of a controller doesn't mean we shouldn't be able to
return a 400 error and run its render() instead of being forced into
its execute().

Signed-off-by: Rick Elrod <rick@elrod.me>